Extracted common arguments
